+++ src/test/java/domain/RacingGameTest.java
@@ -1,4 +1,5 @@
-import domain.Car;
+package domain;
+
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
